[prf/dec][refactor] Restructure prefill-decode ExecutionPlan components to dedicated classes and packages
 Move `LlamaFP16BatchPrefillLayers` to `tornadovm.layers.type.fp16.prefll` and `LlamaFP16FFNLayersForUnifiedDecode` to `tornadovm.layers.type.fp16.decode`
